Refactor CreativeWork and Thing classes to extend BaseType and implement ThingsTrait for improved property handling
